refactor(projects-more): move more projects folder under projects page
